+++ apps/ledger-live-mobile/src/reducers/settings.js
@@ -76,6 +76,7 @@ export type SettingsState = {
   hasCompletedOnboarding: boolean,
   hasInstalledAnyApp: boolean,
   readOnlyModeEnabled: boolean,
+  hasOrderedNano: boolean,
   experimentalUSBEnabled: boolean,
   countervalueFirst: boolean,
   graphCountervalueFirst: boolean,
@@ -117,7 +118,9 @@ export const INITIAL_STATE: SettingsState = {
   orderAccounts: "balance|desc",
   hasCompletedOnboarding: false,
   hasInstalledAnyApp: true,
-  readOnlyModeEnabled: !Config.DISABLE_READ_ONLY,
+  // readOnlyModeEnabled: !Config.DISABLE_READ_ONLY,
+  readOnlyModeEnabled: true,
+  hasOrderedNano: false,
   experimentalUSBEnabled: false,
   countervalueFirst: true,
   graphCountervalueFirst: true,
@@ -411,6 +414,10 @@ const handlers: Object = {
     ...state,
     lastConnectedDevice,
   }),
+  SET_HAS_ORDERED_NANO: (state, action) => ({
+    ...state,
+    hasOrderedNano: action.enabled,
+  }),
   SET_MARKET_REQUEST_PARAMS: (state: SettingsState, { payload }) => ({
     ...state,
     marketRequestParams: {
@@ -619,6 +626,9 @@ export const starredMarketCoinsSelector = (state: State) =>
 
 export const lastConnectedDeviceSelector = (state: State) =>
   state.settings.lastConnectedDevice;
+
+export const hasOrderedNanoSelector = (state: State) =>
+  state.settings.hasOrderedNano;
 
 export const marketRequestParamsSelector = (state: State) =>
   state.settings.marketRequestParams;