+++ apps/ledger-live-mobile/src/screens/Lending/EnableFlow/01-SelectAccount.js
@@ -8,7 +8,6 @@ import { Trans } from "react-i18next";
 import type {
   Account,
   AccountLike,
-  CryptoCurrency,
   TokenCurrency,
 } from "@ledgerhq/live-common/lib/types";
 
@@ -49,7 +48,7 @@ const SEARCH_KEYS = [
 
 type Props = {
   navigation: any,
-  route: { params?: { currency: CryptoCurrency | TokenCurrency } },
+  route: { params?: { token: TokenCurrency } },
 };
 
 const keyExtractor = item => item.account.id;
@@ -57,8 +56,8 @@ const keyExtractor = item => item.account.id;
 function LendingEnableSelectAccount({ route, navigation }: Props) {
   const { colors } = useTheme();
   const locale = useSelector(localeSelector);
-  const currency = route?.params?.currency;
-  invariant(currency, "currency required");
+  const token = route?.params?.token;
+  invariant(token, "token required");
 
   let enabledTotalAmount = null;
   const accounts = useSelector(
@@ -74,10 +73,10 @@ function LendingEnableSelectAccount({ route, navigation }: Props) {
       const n = navigation.getParent() || navigation;
       n.replace(NavigatorName.AddAccounts, {
         screen: ScreenName.AddAccountsTokenCurrencyDisclaimer,
-        params: { token: currency },
+        params: { token },
       });
     }
-  }, [currency, filteredAccounts, navigation]);
+  }, [token, filteredAccounts, navigation]);
 
   let accountsWithUnlimitedEnabledAmount = 0;
   filteredAccounts.forEach(({ account }) => {
@@ -103,7 +102,7 @@ function LendingEnableSelectAccount({ route, navigation }: Props) {
 
   const formattedEnabledAmount =
     enabledTotalAmount instanceof BigNumber &&
-    formatCurrencyUnit(currency.units[0], enabledTotalAmount, {
+    formatCurrencyUnit(token.units[0], enabledTotalAmount, {
       showCode: true,
       disableRounding: false,
       locale,
@@ -235,7 +234,7 @@ function LendingEnableSelectAccount({ route, navigation }: Props) {
       <TrackScreen
         category="Lend Approve"
         name="Select Account"
-        eventProperties={{ currencyName: currency.name }}
+        eventProperties={{ currencyName: token.name }}
       />
       <LendingWarnings />
       <ConfirmationModal
@@ -284,7 +283,7 @@ function LendingEnableSelectAccount({ route, navigation }: Props) {
                       ? filteredAccounts.length
                       : accountsWithUnlimitedEnabledAmount,
                   amount: formattedEnabledAmount,
-                  currency: currency.name,
+                  currency: token.name,
                 }}
                 count={
                   enabledTotalAmount < Infinity