+++ libs/ledger-live-common/src/families/ethereum/hw-signMessage.ts
@@ -2,6 +2,7 @@ import Eth from "@ledgerhq/hw-app-eth";
 import Transport from "@ledgerhq/hw-transport";
 import { TypedDataUtils } from "eth-sig-util";
 import { bufferToHex } from "ethereumjs-util";
+import { getEnv } from "../../env";
 import type { MessageData, Result } from "../../hw/signMessage/types";
 import type { TypedMessageData, TypedMessage } from "./types";
 type EthResolver = (
@@ -36,7 +37,7 @@ const resolver: EthResolver = async (
   if (typeof message === "string") {
     result = await eth.signPersonalMessage(path, rawMessage.slice(2));
   } else {
-    if (process.env.EXPERIMENTAL_EIP712) {
+    if (getEnv("EXPERIMENTAL_EIP712")) {
       result = await eth.signEIP712Message(path, message);
     } else {
       result = await eth.signEIP712HashedMessage(